+++ .mypy-strict.ini
@@ -33,7 +33,7 @@ ignore_errors = True
 ignore_errors = True
 ignore_missing_imports = True
 
-[mypy-job_runner.*]
+[mypy-ert.job_runner.*]
 ignore_errors = True
 
 [mypy-numpy.*]
