+++ .mypy.ini
@@ -29,7 +29,7 @@ ignore_errors = False
 ignore_missing_imports = False
 ignore_errors = False
 
-[mypy-job_runner.*]
+[mypy-ert.job_runner.*]
 ignore_missing_imports = True
 ignore_errors = True
