+++ src/ert/job_runner/job_dispatch.py
@@ -2,7 +2,7 @@ import os
 import signal
 import sys
 
-from job_runner.cli import main as job_runner_main
+from ert.job_runner.cli import main as job_runner_main
 
 
 def sigterm_handler(_signo, _stack_frame):
