+++ src/ert/job_runner/util/data.py
@@ -1,6 +1,6 @@
 """Utility to compensate for a weak job type."""
 import time
-from job_runner.reporting.message import _JOB_STATUS_WAITING
+from ert.job_runner.reporting.message import _JOB_STATUS_WAITING
 
 
 def create_job_dict(job):
