+++ src/res/job_queue/forward_model_status.py
@@ -18,7 +18,7 @@ import json
 import os.path
 import time
 
-from job_runner import JOBS_FILE, STATUS_json
+from ert.job_runner import JOBS_FILE, STATUS_json
 
 
 def _serialize_date(dt):
